requirements.txt
asgiref==3.7.2
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
certifi==2023.7.22
cffi==1.15.1
channels==4.0.0
channels-redis==4.1.0
charset-normalizer==3.2.0
config==0.5.1
constantly==15.1.0
cryptography==41.0.3
daphne==4.0.0
Django==4.2.4
django-restframework==0.0.1
djangorestframework==3.14.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
msgpack==1.0.5
passlib==1.7.4
pexpect==4.8.0
Pillow==10.0.0
ptyprocess==0.7.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.8.0
pyOpenSSL==23.2.0
python-decouple==3.8
pytz==2023.3
redis==5.0.0
requests==2.31.0
service-identity==23.1.0
simplejwt==2.0.1
six==1.16.0
sqlparse==0.4.4
Strip==0.0.1.dev14
stripe==5.5.0
Twisted==22.10.0
twisted-iocpsupport==1.0.4
txaio==23.1.1
typing==3.7.4.3
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
zope.interface==6.0
